Avoid undefined behaviour in the AES CTR counter byte-swap macro
BE_TO_UINT32 shifted a uint8_t, promoted to int, left by 24 bits. When
the top bit of that byte is set the result does not fit in an int, which
is undefined behaviour; UBSan reports it for any 16-byte IV whose byte
12 is 0x80 or above. Cast each byte to uint32_t before shifting.

The macro is used by the ARMv8 CTR paths. The copy in aes256_ni.c is
unused but is kept identical to the other two.

+++ src/common/aes/aes128_armv8.c
@@ -136,7 +136,7 @@ static uint32_t UINT32_TO_BE(const uint32_t x) {
 	y.bytes[3] = x & 0xFF;
 	return y.val;
 }
-#define BE_TO_UINT32(n) (uint32_t)((((uint8_t *) &(n))[0] << 24) | (((uint8_t *) &(n))[1] << 16) | (((uint8_t *) &(n))[2] << 8) | (((uint8_t *) &(n))[3] << 0))
+#define BE_TO_UINT32(n) (((uint32_t)((uint8_t *) &(n))[0] << 24) | ((uint32_t)((uint8_t *) &(n))[1] << 16) | ((uint32_t)((uint8_t *) &(n))[2] << 8) | ((uint32_t)((uint8_t *) &(n))[3] << 0))
 
 
 void oqs_aes128_ctr_enc_sch_upd_blks_armv8(void *schedule, uint8_t *out, size_t out_blks) {
